Add back-off/retry logic to docker ping.
We have a dependency on docker socket being ready to listen to requests for
any of the ecs-init actions to complete. However, on amazon linux, the docker
init script doesn't wait for the socket file to be ready and assumes that the
docker daemon is running when the pid file is created. This results in
ecs-init's pre-stat and start failing as there's a race for the docker socket
file to be ready.

This commit adds a back-off/retry policy around the docker Ping API to remediate
this issue.