Use -e flag when disabling ipv6 router advertisements
This is a follow-up to #313

Some customers might have already disabled ipv6 entirely, which would
mean that the ipv6 accept_ra sysctl key wouldn't exist. In this case,
the sysctl command would fail and the agent would exit.

The -e flag tells sysctl not to fail if the config key doesnt exist.
from sysctl man page:
-e
    Use this option to ignore errors about unknown keys.