packaging/generic-deb: use agent version from file instead of tag.
When we build a new version of ecs init, the new agent version is available from file ecs-init/ECSVERSION, but not available from tag because the tag is created when we publish the release, but we build the package before that. Therefore, use the agent version from the ECSVERSION file instead of from `git describe --tags`.